requirements.txt
asgiref==3.8.1
cloudinary==1.40.0
dj-database-url==2.2.0
dj-rest-auth==6.0.0
Django==5.0.6
django-allauth==0.63.3
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.5.0
